+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [1.2.0] - 2022-10-04
+
+### Added
+
+- TwoFactorAuth class and module.
+- MessageService class.
+- Documentation in russian.
+
+### Changed
+
+- Password now checking for minimum and maximum length when created.
+- Password now checking against top 1000 most popular password when created.
+- User's account now can be locked for a wile after several faild login attempts.
+
 ## [1.0.1] - 2022-09-20
 
 ### Changed
@@ -22,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Security class for password encoding and decoding.
 - Schema and types for Authentication module.
 
-[unreleased]: https://github.com/web-soft-llc/web-soft-server/compare/v1.0.1...master
+[unreleased]: https://github.com/web-soft-llc/web-soft-server/compare/v1.2.0...master
+[1.2.0]: https://github.com/web-soft-llc/web-soft-server/releases/tag/v1.0.1...v1.2.0
 [1.0.1]: https://github.com/web-soft-llc/web-soft-server/releases/tag/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/web-soft-llc/web-soft-server/releases/tag/v1.0.0